Modified Dockerfile to build armv7 images on x86 hosts
This was done by using the documentation and images of the good
people at resin.io.

Unfortunately the docker image diverges quite a bit in its essence,
because it would be downloading binaries for the newest build of
MariaDB, instead of building it from source. These builds are not
supplied in ARM, so we will take the older builds from the Debian
Jessie ARM repository.